# Speed up iteration with numbers