--- a/base/number.jl
+++ b/base/number.jl
@@ -48,9 +48,9 @@ angle(z::Real) = atan2(zero(z), z)
 
 widemul(x::Number, y::Number) = widen(x)*widen(y)
 
-start(x::Number) = false
-next(x::Number, state) = (x, true)
-done(x::Number, state) = state
+start(x::Number) = 0  # see #16687
+next(x::Number, state) = (x, state+1)
+done(x::Number, state) = state == 1
 isempty(x::Number) = false
 in(x::Number, y::Number) = x == y