@Jutho 2016-05-31
Might this be a regression? It doesn't happen in v0.4.5.