@vtjnash 2016-06-01
> Might this be a regression? It doesn't happen in v0.4.5.

Yes, this appears to be an LLM regression since it also doesn't happen with llvm 3.3 on master