@timholy 2016-06-01
How does one go about reporting such things upstream? I'm presuming that a few lines of Julia code that demonstrate the problem won't quite cut it. Since the code returned by `@code_llvm` appears to be something that has already passed through an optimizer, is there a good way to capture the initial input?