@yuyichao 2016-06-01
Run it with `-O0`.