@vtjnash 2016-06-01
Usually upstream wants a .ll file:
`open("code.ll", "w") do io; code_llvm(io, f, args, #=strip=#false, #=module=#true); end`

Which you can then test outside of julia with `llc` (from julia/usr/bin) and look at the effects of various optimizations levels (`-O1/2/3`), etc. on the resulting assembly and intermediate IR (`-print-after-all`)