@timholy 2016-06-01
Thanks for the great tips. I wrote the `.ll` files using -O0 as a julia option. But in playing with `llc`, I could neither see any difference in the resulting assembly depending on the optimization level (above -O0), nor (assuming I'm reading this correctly) get the "fast" version to elide the loop.

Examples:
```
tim@diva:/tmp$ ~/src/julia-0.5/usr/bin/llc -O3 -o fast_3 -print-after-all loop_fast.ll 2>intermed_fast_3
tim@diva:/tmp$ ~/src/julia-0.5/usr/bin/llc -O1 -o fast_1 -print-after-all loop_fast.ll 2>intermed_fast_1
tim@diva:/tmp$ cmp fast_1 fast_3
tim@diva:/tmp$ cmp intermed_1 intermed_3
```

I posted a gist with `intermed_fast_2` and `intermed_slow_2` [here](https://gist.github.com/timholy/40f99ce28d431e5f24113bd495f0b90f)

I'm sure I'm being a noob about this, so apologies in advance. But I'm also wondering, are we certain this issue is purely an LLVM issue? Or is there some pass I have to turn on explicitly in `llc`?