@tkelman 2016-06-04
This needs to be debugged and isolated a bit more.