@timholy 2016-06-04
I agree, I'll file an issue. But it's too separate, with too simple of a workaround, to let it derail my current task.