@tkelman 2016-06-04
yeah that's fine. workaround was simple in this case that you noticed, but how much other code might be affected by the same underlying problem?