@timholy 2016-06-04
I'm not disagreeing in the slightest (it's much of why I filed the issue).